refactor: rename theme module to theme_plotly for library-specific theming
Module Renaming:
- Rename theme.py → theme_plotly.py (preserves git history)
- Rename test_theme.py → test_theme_plotly.py

Import Updates:
- Update chart_evolution.py, chart_evolution_commit.py, chart_breakdown.py
- Update test imports with multi-line formatting

Prepares codebase for Altair chart support by making Plotly-specific theming
explicit. Enables future theme_altair.py module with clear separation between
charting libraries.
