+++ src/plot_py_repo/chart_evolution_commit.py
@@ -7,7 +7,7 @@ import numpy as np
 import pandas as pd
 import plotly.express as px
 
-from .theme import add_footnote_annotation, apply_common_layout, save_chart_image
+from .theme_plotly import add_footnote_annotation, apply_common_layout, save_chart_image
 
 CHART_TITLE = "Repository Growth by Commit"
